+++ spi/spi_master_slave/spi_master/utilities/hardcoded_functions.cpp
@@ -3,7 +3,7 @@
 #include "../transceiver/rx_core.hpp"
 #include <pico/multicore.h>
 
-void set_freq(uint16_t freq)
+void set_freq(uint32_t freq)
 {
   current_freq = freq;
